Remove unused code and refactor code structure
