+++ src/core/Common/InstructionFormats.ts
@@ -12,38 +12,3 @@ export enum Formats {
 }
 
 export let FormatsNames: string[] = ["TwoGeneralRegisters", "TwoFloatingRegisters", "GeneralRegisterAndInmediate", "GeneralLoadStore", "FloatingLoadStore", "Jump", "Noop"];
-
-export function opcodeToFormat(opcode: Opcodes): Formats {
-    switch (opcode) {
-        case Opcodes.NOP:
-            return Formats.Noop;
-        case Opcodes.ADD:
-        case Opcodes.SUB:
-        case Opcodes.MULT:
-        case Opcodes.OR:
-        case Opcodes.AND:
-        case Opcodes.XOR:
-        case Opcodes.NOR:
-        case Opcodes.SLLV:
-        case Opcodes.SRLV:
-            return Formats.TwoGeneralRegisters;
-        case Opcodes.ADDF:
-        case Opcodes.SUBF:
-        case Opcodes.MULTF:
-            return Formats.TwoFloatingRegisters;
-        case Opcodes.ADDI:
-            return Formats.GeneralRegisterAndInmediate;
-        case Opcodes.SW:
-        case Opcodes.LW:
-            return Formats.GeneralLoadStore;
-        case Opcodes.SF:
-        case Opcodes.LF:
-            return Formats.FloatingLoadStore;
-        case Opcodes.BNE:
-        case Opcodes.BEQ:
-        case Opcodes.BGT:
-            return Formats.Jump;
-        default:
-            throw new Error('Invalid opcode: ' + opcode);
-    }
-}
