--- src/core/Common/MachineStatus.ts
@@ -1,5 +1,0 @@
-export interface MachineStatus {
-    cycle: number;
-    executing: boolean;
-    breakPoint: boolean;
-}
