+++ src/core/Common/Opcodes.ts
@@ -1,4 +1,5 @@
 import { FunctionalUnitType } from "./FunctionalUnit";
+import { Formats } from "./InstructionFormats";
 
 export enum Opcodes {
     NOP = 0,
@@ -58,5 +59,40 @@ export function opcodeToFunctionalUnit(opcode: Opcodes): FunctionalUnitType {
             return FunctionalUnitType.JUMP;
         default:
             throw new Error("Error at opcodeToFunctionalUnit, unknown opcode : " + OpcodesNames[opcode]);
+    }
+}
+
+export function opcodeToFormat(opcode: Opcodes): Formats {
+    switch (opcode) {
+        case Opcodes.NOP:
+            return Formats.Noop;
+        case Opcodes.ADD:
+        case Opcodes.SUB:
+        case Opcodes.MULT:
+        case Opcodes.OR:
+        case Opcodes.AND:
+        case Opcodes.XOR:
+        case Opcodes.NOR:
+        case Opcodes.SLLV:
+        case Opcodes.SRLV:
+            return Formats.TwoGeneralRegisters;
+        case Opcodes.ADDF:
+        case Opcodes.SUBF:
+        case Opcodes.MULTF:
+            return Formats.TwoFloatingRegisters;
+        case Opcodes.ADDI:
+            return Formats.GeneralRegisterAndInmediate;
+        case Opcodes.SW:
+        case Opcodes.LW:
+            return Formats.GeneralLoadStore;
+        case Opcodes.SF:
+        case Opcodes.LF:
+            return Formats.FloatingLoadStore;
+        case Opcodes.BNE:
+        case Opcodes.BEQ:
+        case Opcodes.BGT:
+            return Formats.Jump;
+        default:
+            throw new Error('Invalid opcode: ' + opcode);
     }
 }
