+++ src/core/VLIW/VLIWOperation.ts
@@ -82,10 +82,4 @@ export class VLIWOperation extends Instruction {
     public setPredFalse(p: number) {
         this._predicateFalse = p;
     }
-
-    // TODO: remove this, use Instruction.isJump instead
-    public isJump() {
-        return (this._opcode === Opcodes.BEQ) || (this._opcode === Opcodes.BGT) || (this._opcode === Opcodes.BNE);
-    }
-
 }
