--- src/core/VLIW/index.ts
@@ -1,3 +1,0 @@
-export * from './VLIWError';
-export * from './VLIWCode';
-export * from './VLIW';
